store: move StoreOpener to its own module (#7559)
Just a tiny bit of refactoring separating configuration part of the
storage and code responsible for opening the storage into separate
modules.
